requirements.txt
Pillow>=8.0.0
streamdeck
urllib3
svglib
websocket-client
cairosvg